refactor: move common modal components into their own subfolder
