+++ dev-client/src/navigation/navigators/BottomNavigator.tsx
@@ -17,7 +17,7 @@
 import {memo, useCallback, useEffect} from 'react';
 import {useDispatch, useSelector} from 'terraso-mobile-client/store';
 import {useTranslation} from 'react-i18next';
-import {ConfirmModal} from 'terraso-mobile-client/components/ConfirmModal';
+import {ConfirmModal} from 'terraso-mobile-client/components/modals/ConfirmModal';
 import {signOut} from 'terraso-client-shared/account/accountSlice';
 import {BottomNavIconButton} from 'terraso-mobile-client/navigation/components/BottomNavIconButton';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
